TDS-604: OnPaused/OnCompleted implementation

+++ service/src/main/java/tds/exam/models/FieldTestItemGroup.java
@@ -176,7 +176,7 @@ public class FieldTestItemGroup {
         private Integer positionAdministered;
         private Instant administeredAt;
 
-        public Builder withId(int id) {
+        public Builder withId(long id) {
             this.id = id;
             return this;
         }
